backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
mysqlclient
dotenv
dj_database_url
whitenoise